test: split the suite into unit and integration
tests/unit is everything with synthetic fixtures; tests/integration is
the two modules that shell out to a real external tool (uv build,
git-cliff) and self-skip when it is missing, so `pytest -m "not
integration"` skips them outright.

Moved with git mv so history and blame follow each file. conftest.py
stays at tests/ top level so it applies to both subdirectories.
