+++ tests/unit/test_gpu_detect.py
