+++ tests/unit/test_image_store.py
