+++ tests/unit/test_installer_scripts.py
