+++ tests/unit/test_model_images.py
