+++ tests/unit/test_package_mode.py
