+++ tests/unit/test_paths.py
