+++ tests/unit/test_serial_emulator.py
