+++ tests/unit/test_slot_details.py
