+++ tests/unit/test_slot_templates.py
