+++ tests/unit/test_theme_editor.py
