+++ tests/unit/test_training_manager.py
